refactor: move version-related function to separate file
